backend/requirements.txt
alembic==1.10.1
APScheduler==3.10.0
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
PyJWT==2.6.0
PyMySQL==1.0.2
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
SQLAlchemy==2.0.4
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
uvicorn==0.21.1
Werkzeug==2.2.3
